pyproject.toml
[project]
authors = [
    {name = "Census", email = "dev@getcensus.com"},
]
license = "MIT"
license-files = ["LICENSE"]
requires-python = "<4.0,>=3.10"
dependencies = [
    "apache-airflow>=2.10,<4",
    "apache-airflow-providers-common-compat>=1.8.0",
    "apache-airflow-providers-http>=1.1.1",
]
name = "airflow-provider-census"
version = "2.1.0"
description = "A Census provider for Apache Airflow."
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]

[project.urls]
"Bug Tracker" = "https://github.com/sutrolabs/airflow-provider-census/issues"
homepage = "https://www.getcensus.com/"
repository = "https://github.com/sutrolabs/airflow-provider-census"

[project.entry-points.apache_airflow_provider]
provider_info = "airflow_provider_census.__init__:get_provider_info"

[dependency-groups]
dev = [
    "pytest<7.0.0,>=6.2.4",
    "requests-mock<2.0.0,>=1.9.3",
]
